Add a new group of users: teachers.
Add a teacher menu and create the teacher directory with
stub files to link to from the menu.

In order to reuse the attendance_list report from the
admin pages, test for type of user to show the correct menu.

Rename base_student.html to base_user.html because it
is now used by students and teachers.

Add a little information on the welcome page for teachers.